86: match float64 sampler output to float32 ground truth
heun_sample integrates in double on purpose (gaussian_diffusion.py, "Use float64
for accuracy"), so samples and samples_act arrive as Double while the GT multi-hot
and S3D's weights are float32. Two faults, both only reachable on real tensors:

  binary_cross_entropy demands an exact dtype match rather than promoting, so the
  raw and raw_fused CE path raised and took down the whole validation pass.
  conv3d likewise refused the Double generated half, so FVD logged nothing --
  caught by the try/except, which is exactly the silent skip it was there for.

The harness stub returned float32 and so could not see either. It now returns
double like the real sampler; all six of its tests fail without these casts.

+++ improved_diffusion/action_ce.py
@@ -21,6 +21,9 @@ EPS = 1e-6
 
 def keypress_cross_entropy(p, y):
     """Bernoulli CE of probabilities p against multi-hot y, summed over keys, meaned over frames."""
+    # heun_sample integrates in float64, so p arrives Double while the GT multi-hot is
+    # float32; binary_cross_entropy demands an exact match rather than promoting.
+    y = y.to(p.dtype)
     # clamp to EPS so that if model predicts 0 or 1, we don't get log(0) errors.
     p = p.clamp(EPS, 1 - EPS)
     return F.binary_cross_entropy(p, y, reduction="none").sum(dim=-1).mean()
@@ -37,7 +40,7 @@ def keypress_ce_baserate(y, q=None):
     read this as an anchor to beat, not as the policy's press rate."""
     if q is None:
         q = y.reshape(-1, y.shape[-1]).mean(dim=0)
-    q = q.clamp(0.0, 1.0)
+    q = q.to(y.dtype).clamp(0.0, 1.0)
     # xlogy is used to guard against log(0) errors.
     per_frame = -(torch.special.xlogy(y, q) + torch.special.xlogy(1 - y, 1 - q)).sum(dim=-1)
     return per_frame.mean()
